Fix bug on move copy, reset package level
On stock.move, the `package_level_id` field did not have a `copy=False`
attribute when version 13.0 of Odoo was released. So when we create
a new move, the package level of the move being copied was linked
as well to the new move.

It has been fixed recently in odoo in
https://github.com/odoo/odoo/commit/ecf726ae8221e6871c1e391294c633d8b6bcaa9a
but to be on the safe side if the code is not up-to-date,
it's anyway better to force a False value.