[ADD] stock_measuring_device_zippcube[_reception_screen]
On-going development. First 'clean' version, not yet
preared to be refactored.

Re-does the original module stock_zipcube, proposed in the PR
https://github.com/OCA/stock-logistics-warehouse/pull/1096
so that it doesn't depend on stock_cubiscan. For the moment,
stock_measuring_device_zippcube depends on itself, but the idea
is to extract the common parts for other measuring devices,
such as Cubiscan, and put them in some other module.

The measuring has been made available from the reception screen,
and this is implemented into the new module with name
stock_measuring_device_zippcube_reception_screen, that maybe
could go into the repository wms, since reception_screen is
from there.